+++ plone/app/users/browser/userdatapanel.py
@@ -74,8 +74,6 @@ class UserDataPanelAdapter(AccountPanelSchemaAdapter):
 
 class UserDataPanel(AccountPanelForm):
 
-    label = _(u'title_personal_information_form',
-              default=u'Personal Information')
     form_name = _(u'User Data Form')
     enableCSRFProtection = True