+++ CHANGES.rst
@@ -4,6 +4,9 @@ CHANGES
 2.2 (unreleased)
 ----------------
 
+- Adjust navigation markup for Plone 5.
+  [davisagli]
+
 - Use email_from_address from registry (Plone 5) in tests.
   [khink]