+++ CHANGES.rst
@@ -4,6 +4,9 @@ CHANGES
 2.2 (unreleased)
 ----------------
 
+- Read security settings from new Plone 5 registry.
+  [jure]
+
 - Ported tests to plone.app.testing
   [gforcada, tomgross]
 
@@ -12,6 +15,7 @@ CHANGES
 
 - Use email_from_address from registry (Plone 5) in tests.
   [khink]
+
 
 2.1.0 (2014-10-23)
 ------------------